+++ data/src/main/java/com/dida/data/model/dex/PostSellNftRequest.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class PostSellNftRequest(
     @SerializedName("payPwd") val payPwd: String,
     @SerializedName("nftId") val nftId: Long,
-    @SerializedName("price") val price: Float,
+    @SerializedName("price") val price: Double,
 )
